refactor home, component
