+++ lib/app/pages/home/home_view.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_rounded_date_picker/flutter_rounded_date_picker.dart';
-import 'package:flutter_svg/svg.dart';
-import 'package:intl/intl.dart';
 import 'package:percent_indicator/circular_percent_indicator.dart';
 import 'package:personal_financial_management/app/components/colors/my_colors.dart';
-import 'package:personal_financial_management/app/components/date_picker/date_picker.dart';
+import 'package:personal_financial_management/app/components/date_picker/date_controller.dart';
 import 'package:personal_financial_management/app/components/icons/my_icons.dart';
-import 'package:personal_financial_management/app/components/images/my_images.dart';
+import 'package:personal_financial_management/app/utils/utils.dart';
 
 class HomeView extends StatefulWidget {
   const HomeView({Key? key}) : super(key: key);
@@ -19,6 +16,80 @@ class _HomeViewState extends State<HomeView> {
   DateTime? dateTime;
   @override
   Widget build(BuildContext context) {
+    return _buildTabBar();
+  }
+
+  // Widgets
+  // TabBar
+  Widget _buildTabBar() {
+    late int _currentIndex = 0;
+    late TextStyle _tabBarTextStyle = TextStyle(
+      fontSize: 16,
+    );
+    late List<Widget> _tabs = [
+      Tab(
+        child: Text(
+          'THÁNG',
+          style: _tabBarTextStyle,
+        ),
+      ),
+      Tab(
+        child: Text(
+          'TUẦN',
+          style: _tabBarTextStyle,
+        ),
+      ),
+      Tab(
+        child: Text(
+          'NGÀY',
+          style: _tabBarTextStyle,
+        ),
+      ),
+    ];
+    late List<Widget> _tabViews = [
+      _buildMonthTabView(),
+      _buildWeekTabView(),
+      _buildDayTabView(),
+    ];
+    void _onChangeTab(int index) {
+      setState(() {
+        _currentIndex = index;
+      });
+    }
+
+    return DefaultTabController(
+      initialIndex: _currentIndex,
+      length: 3,
+      child: Scaffold(
+          appBar: PreferredSize(
+            preferredSize: const Size.fromHeight(50),
+            child: Container(
+              decoration: BoxDecoration(
+                boxShadow: <BoxShadow>[
+                  BoxShadow(
+                    color: MyAppColors.gray500,
+                    blurRadius: 15.0,
+                    offset: const Offset(0.0, 0.75),
+                  )
+                ],
+                color: MyAppColors.gray050,
+              ),
+              child: TabBar(
+                indicatorColor: MyAppColors.accent700,
+                unselectedLabelColor: MyAppColors.gray600,
+                labelColor: MyAppColors.accent700,
+                onTap: _onChangeTab,
+                tabs: _tabs,
+              ),
+            ),
+          ),
+          body: TabBarView(
+            children: _tabViews,
+          )),
+    );
+  }
+
+  Widget _buildMonthTabView() {
     return WillPopScope(
       onWillPop: () async {
         return false;
@@ -27,7 +98,6 @@ class _HomeViewState extends State<HomeView> {
         color: MyAppColors.white000,
         child: Column(
           children: [
-            _buildTabBar(),
             MyDatePicker(dateTime: dateTime),
             _buildIndicatorChart(),
             Container(
@@ -62,79 +132,16 @@ class _HomeViewState extends State<HomeView> {
     );
   }
 
-  Widget _buildTabBar() {
-    late int _currentIndex = 0;
-    late TextStyle _tabBarTextStyle = TextStyle(
-      fontSize: 16,
-    );
-    late List<Widget> _tabs = [
-      Tab(
-        child: Text(
-          'THÁNG',
-          style: _tabBarTextStyle,
-        ),
-      ),
-      Tab(
-        child: Text(
-          'TUẦN',
-          style: _tabBarTextStyle,
-        ),
-      ),
-      Tab(
-        child: Text(
-          'NGÀY',
-          style: _tabBarTextStyle,
-        ),
-      ),
-    ];
-    late List<Widget> _tabViews = [];
-    void _onChangeTab(int index) {
-      setState(() {
-        _currentIndex = index;
-      });
-    }
-
-    return Card(
-      color: MyAppColors.gray050,
-      elevation: 5,
-      margin: const EdgeInsets.fromLTRB(0.0, 0.0, 0.0, 8.0),
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(0.0),
-      ),
-      child: DefaultTabController(
-        initialIndex: _currentIndex,
-        length: 3,
-        child: Container(
-          decoration: BoxDecoration(
-            boxShadow: <BoxShadow>[
-              BoxShadow(
-                color: MyAppColors.gray500,
-                blurRadius: 15.0,
-                offset: const Offset(0.0, 0.75),
-              )
-            ],
-            color: MyAppColors.gray050,
-          ),
-          child: TabBar(
-            indicatorColor: MyAppColors.accent700,
-            unselectedLabelColor: MyAppColors.gray600,
-            labelColor: MyAppColors.accent700,
-            onTap: _onChangeTab,
-            tabs: _tabs,
-          ),
-        ),
-      ),
-    );
+  Widget _buildWeekTabView() {
+    return Container();
   }
 
+  Widget _buildDayTabView() {
+    return Container();
+  }
+
+  // Chart indicator
   Widget _buildIndicatorChart() {
-    //format number according to #.##0.00 format
-    NumberFormat _numberFormat = NumberFormat.currency(
-      locale: 'vi-VN',
-      symbol: '',
-      decimalDigits: 2,
-      customPattern: '#,##0.00',
-    );
     return CircularPercentIndicator(
       addAutomaticKeepAlive: true,
       reverse: true,
@@ -158,7 +165,7 @@ class _HomeViewState extends State<HomeView> {
           Padding(
             padding: const EdgeInsets.all(14.0),
             child: Text(
-              '${_numberFormat.format(123459123)} ${_numberFormat.currencyName}',
+              '${numberFormat.format(123459123)} ${numberFormat.currencyName}',
               textAlign: TextAlign.center,
               style: TextStyle(
                 fontSize: 30,
@@ -173,7 +180,7 @@ class _HomeViewState extends State<HomeView> {
               maxWidth: 200,
             ),
             child: Text(
-              'Hạn mức: ${_numberFormat.format(200000000)} ${_numberFormat.currencyName}',
+              'Hạn mức: ${numberFormat.format(200000000)} ${numberFormat.currencyName}',
               maxLines: 1,
               textAlign: TextAlign.center,
               style: TextStyle(
@@ -192,6 +199,7 @@ class _HomeViewState extends State<HomeView> {
     );
   }
 
+  // History Expense
   Widget _buildHistoryTitle() {
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 16.0),
@@ -220,4 +228,7 @@ class _HomeViewState extends State<HomeView> {
       ),
     );
   }
+
+  //Handler for widgets
+
 }
